// states/state.interface.ts
export interface ApplicationData {
  date: Date;
  binNumber: string;
  commissioneRate: string;
  division: string;
  subDivision: string;
  circle: string;
  shopName: string;
  brandName: string;
  areaOrshoppingMall: string;
  businessRegisteredAddress: string;
  outletAddress: string;
  nid: string;
  name: string;
  mobile: string;
  email: string;
  startDate: Date;
  category: string;
  mushak: string;
  subCategory: string;
  numberOfOutlet: string;
  numberOfCounter: string;
  differentBin: string;
  transactionType: string;
  posSoftwareProvider: string;
  nrbApproved: string;
  thirdPartyName: string;
  monthlyAverageSales: string;
  monthlyAverageCustomer: string;
  onlineSaleAvailable: string;
  onlineSaleParcent: string;
  onlineOrderMode: string;
  productInfo: string;
  productName: string;
  productUnit: string;
  unitPrice: string;
  vatParcent: string;
  sdPercent: string;
  priceIncludingVat: string;
  priceExcludingVat: string;
  stockKeeping: string;
  posSoftware: string;
  posPrinter: string;
  pcOrLaptop: string;
  router: string;
  networking: string;
  surveillance: string;
  mobileOperator: string;
  operatorCoverage: string;
  shopPic: HTMLImageElement | null;
  binCertificate: HTMLImageElement | null;
  itemList: HTMLImageElement[];
}